+++ app/frontend/features/AccountManager/AccountForm/AccountForm.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Decorator, Meta, StoryObj } from '@storybook/react'
-import AccountForm from './AccountForm'
+import AccountForm from '../AccountForm'
 import AccountProvider from '../AccountProvider'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import LogTransportProvider from '@/components/LogTransportProvider'
@@ -20,8 +20,14 @@ const accountProviderDecorator: Decorator = (Story, { args }) => (
   </LogTransportProvider>
 )
 
+const accountDataDecorator: Decorator = (Story, { args }) => {
+  console.debug('Loading mock account data...')
+
+  return <Story {...args} />
+}
+
 const meta: Meta<typeof AccountForm> = {
-  decorators: [accountProviderDecorator],
+  decorators: [accountDataDecorator, accountProviderDecorator],
   title: 'AccountManager/AccountForm',
   component: AccountForm,
 }
@@ -32,4 +38,22 @@ type Story = StoryObj<typeof meta>
 
 export const Default: Story = {
   args: { readOnly: true, compact: false },
+}
+
+/**
+ * This story demonstrates how to load a Business account. The account
+ * data is only updated when the `accountId` prop changes while the form
+ * is in read-only mode.
+ */
+export const BusinessForm: Story = {
+  args: { initialType: 'Business', accountId: '4321', readOnly: true },
+}
+
+/**
+ * This story demonstrates how to load an Individual account. The account
+ * data is only updated when the `accountId` prop changes while the form
+ * is in read-only mode.
+ */
+export const IndividualForm: Story = {
+  args: { initialType: 'Individual', accountId: '1234', readOnly: true },
 }
