+++ app/frontend/features/AccountManager/AccountSummaryModal.tsx
@@ -4,7 +4,7 @@ import withAllTheProviders from '@/components/withAllTheProviders'
 import LoadingAnimation from '../../components/LoadingAnimation'
 import { useAccountContext, withAccountProvider } from '@/features/AccountManager/AccountProvider'
 import AccountTitleLabel from './AccountTitleLabel'
-import AccountForm from './AccountFormWithFormik'
+import AccountForm from './AccountForm'
 import CloseIcon from '@/components/Icons/CloseIcon'
 import AccountSlug from './AccountSlug'
 import { useLogTransport } from '@/components/LogTransportProvider'
